Move TextJustify to its own file.
When we generate external types like Length in ComputedStyle, we specify
the path to the type (e.g. "platform/Length"). We then deduce the name
of the type from the path ("Length"). This means that the type must be
stored in a file with the same name.

This patch moves TextJustify to its own file so that it
can be generated.

BUG=628043

Review-Url: https://codereview.chromium.org/2883813002
Cr-Original-Commit-Position: refs/heads/master@{#471948}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dc8b3ef5667dfec73a697d1115e19832ea6341bb