Upstream service worker `ready` test to WPT
This test exists in both WPT and the Chromium project tree. Although the
two implementations assert the same behavior, the Chromium version is
preferable for a number of reasons:

- It uses the `promise_test` function and expresses the test body in
  terms of promises, resulting in improved test readability.
- It includes additional sub-tests

Update the upstream version of the test with these improvements. In
addition:

- remove unnecessary precaution from newly-introduced sub-tests
- remove the dynamically-generated iframe at the conclusion of each
  sub-test

Remove the Chromium-specific version of the test.

BUG=688116
R=falken@chromium.org

Review-Url: https://codereview.chromium.org/2880823002
Cr-Original-Commit-Position: refs/heads/master@{#472021}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 981a96bc5df9e0c392193199c5a6d701a973f0f0