Avoid synchronous stylesheet update on html import loaded.
A use counter was updating the styleSheets list, which updated all of
active style to figure out if an html import contains stylesheets.
Instead, do a simpler walk of the stylesheet candidate nodes and return
early if one of them has a sheet, or a sheet load is in progress.

BUG=717506

Review-Url: https://codereview.chromium.org/2882983002
Cr-Original-Commit-Position: refs/heads/master@{#472139}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9d4c419700c0edc7602c27d29a5b520f00bbe13a