[ES6 modules] ModuleMap::GetFetchedModuleScript to return nullptr when entry not found / "fetching"
Before this CL, we asserted that ModuleMap::GetFetchedModuleScript always queried load completed module scripts, but it was not always the case.
This CL changes the member function to return nullptr if the entry doesn't exist or the entry is being fetched.

This CL fixes a crash that happens when we early-exit module script graph fetch by a sub-graph instantiation failed, but there are still partial subgraph in flight.

BUG=718442,594639

Review-Url: https://codereview.chromium.org/2860993002
Cr-Original-Commit-Position: refs/heads/master@{#472192}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1f92cfbdedaecb97e8a87773deb32f98c6997351