Add DocumentMarkerList::MarkerType()
We need a method on DocumentMarkerList that returns the marker type the list
supports so we know when we can safely cast a DocumentMarkerList* to an impl
type.

This CL adds two subclasses of SpellCheckMarkerListImpl, SpellingMarkerListImpl,
and GrammarMarkerListImpl, so we can implement this method.

Note that adding this method also requires modifying
GenericDocumentMarkerListImpl to take a MarkerType param (we're going to
eliminate GenericDocumentMarkerListImpl once we add TextMatchMarkerListImpl but
we need this for now).

BUG=707867

Review-Url: https://codereview.chromium.org/2868413002
Cr-Original-Commit-Position: refs/heads/master@{#472206}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d8b7adbe4b26fedd0a87ffd74aaac141a4694490