Remove references to box_data_ in ComputedStyle.
To allow changes to where fields are stored, ComputedStyle code should
not refer directly to a group, as that code will break when we change
groups. This patch removes references to box_data_ in ComputedStyle,
replacing with generated or handwritten getters. This patch does not
remove references within diffing functions as those will soon be
generated.

BUG=710938

Review-Url: https://codereview.chromium.org/2888513002
Cr-Original-Commit-Position: refs/heads/master@{#472238}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9b38f0b5a429e3046188ae68e65765d57dbe018c