DevTools: add new mode for styles sidebar that hides computed split view
Currently, the default behavior when resizing ElementsPanel will switch from
vertical to horizontal mode when it's width is < 680px. At super small widths,
this makes the `Styles` and `Computed` views both tall and cramped for space.

This CL adds another rule when the width is < 415px, giving the `Computed` view
its own separate tab. A screenshot of DevTools at this width is in the crbug.

BUG=619064

Review-Url: https://codereview.chromium.org/2870163002
Cr-Original-Commit-Position: refs/heads/master@{#472267}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eba1e6e57e86f481a3b5a40713a744ce3d9f2d66