Remove references to background_data_ in ComputedStyle.
To allow changes to where fields are stored, ComputedStyle code should
not refer directly to a group, as that code will break when we change
groups. This patch removes references to background_data_ in ComputedStyle,
replacing with generated or handwritten getters.

References in the cpp file didn't exist.

BUG=710938

Review-Url: https://codereview.chromium.org/2885753002
Cr-Original-Commit-Position: refs/heads/master@{#472364}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0e9b44fea13a90aff88c5a0e7e0164fa99e2418a