Don't show touch handles when selection hidden
If the selection is hidden then clear the touch handles.

Also changed IME test back to using radio input as it was prior to
crrev.com/2616623002. Tapping on a radio button is better to test focus change
because it doesn't create a selection. This means that the clearing of the
handle has to be done by the focus change and not because a new selection was
made.

BUG=714525

Review-Url: https://codereview.chromium.org/2847663002
Cr-Original-Commit-Position: refs/heads/master@{#472371}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 671e6334b0583671f907a52c23a1258b5e1d8770