Move DocumentMarkerList::RemoveMarkersUnderWords() to SpellCheckMarkerListImpl
Currently this is a method on the general DocumentMarkerList interface. This CL
moves it to only be on SpellCheckMarkerListImpl and adds a typecast at the
callsite in DocumentMarkerController to enable this.

I am leaving the method implementation in DocumentMarkerListEditor.cpp in this CL
to make this CL smaller; I will move it to SpellCheckMarkerListImpl.cpp in a future
CL.

BUG=707867

Review-Url: https://codereview.chromium.org/2871823002
Cr-Original-Commit-Position: refs/heads/master@{#472395}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d38becdb2a4dda654d795cb841111778673c0b1d