Remove usage of RefPtr::Release() passed to ComputedStyle setters.
In a follow up patch [1], we remove the use of PassRefPtr in several
ComputedStyle setters by replacing them with RefPtr. However, this means
that callers which use RefPtr::Release to convert a RefPtr into a
PassRefPtr in order to pass it to one of those setters no longer make
sense. This patch simply goes through the callers of these setters and
replaces any RefPtr::Release with std::move to bypass the usage of
PassRefPtrs entirely.

[1] https://codereview.chromium.org/2881143004

BUG=494719

Review-Url: https://codereview.chromium.org/2889903002
Cr-Original-Commit-Position: refs/heads/master@{#472675}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 01bbf1a419789e2f031e00b60f62a2de92e9feb3