Stop retrieving scrolling element id from the CompositorElementId.
Instead, store it locally. This is possible because ScrollState is the only
way to set current_native_scrolling_element on cc::ScrollStateData. Furthermore,
ScrollState is a garbage-collected, script-observable class, so referencing
Element from it is safe.

Also note that ScrollState is part of an unlaunched feature.

BUG=718564

Review-Url: https://codereview.chromium.org/2881423004
Cr-Original-Commit-Position: refs/heads/master@{#472826}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1713618a7ed0036cb64f87a096a94f43a6a8761e