Do not use overridden getter names for internal getters.
Currently, we append the word 'internal' to the name of a field's getter
to form the name of the internal getter. However, there are cases where
this forms a bad name. For example, the getter for 'line-height' is
overridden to be SpecifiedLineHeight, which contains complex logic.
This means that the internal getter will be called
SpecifiedLineHeightInternal. This is bad because the internal getter
does not have any logic to do with specified line height. A more apt
name would be just LineHeightInternal.

This patch changes the internal getters to be <field name> + 'internal'.

Diff of generated files:
https://gist.github.com/780cea9912e024f157a97fba24ee5a0e/revisions

BUG=628043

Review-Url: https://codereview.chromium.org/2890743003
Cr-Original-Commit-Position: refs/heads/master@{#473031}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9f0d5e6a0694f57b9a0df2c0336f506bcf5de9e1