Remove duplicate service worker test
This test exists in both WPT and the Chromium project tree. The WPT
version is slightly weaker because it accounts for a condition that is
not possible according to the latest version of the Service Workers
specification [1].

> # Install
>
> [...]
>
> 16. Run the Update Registration State algorithm passing registration,
>     "waiting" and registration’s installing worker as the arguments.
> 17. Run the Update Registration State algorithm passing registration,
>     "installing" and null as the arguments.
> 18. Run the Update Worker State algorithm passing registration’s
>     waiting worker and installed as the arguments.
> 19. Invoke Finish Job with job.
> 20. Wait for all the tasks queued by Update Worker State invoked in
>     this algorithm have executed.
> 21. Invoke Try Activate with registration.

Due to step 20, worker activation does not occur until all prior tasks
(which includes microtasks such as the Promise handler microtask created
in the test's body) have executed.

Remove the conditional logic from the WPT version and remove the
Chromium test file. Explicitly un-register workers following test
completion.

[1] https://w3c.github.io/ServiceWorker/#installation-algorithm,
    retrieved on 17.05.18

BUG=688116
R=falken@chromium.org

Review-Url: https://codereview.chromium.org/2873373004
Cr-Original-Commit-Position: refs/heads/master@{#473227}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 51eb76c3327468323acfea2393253f846d5aea25