Finish ../platform => public/platform refactor
In https://codereview.chromium.org/2887233003, I changed most of the header
includes in the third_party/WebKit/public/web directory that started with
"../platform/" to instead start with "public/platform/" in order to make
these header includes valid under a DEPS rule that's flagging new includes
of the form "../platform/" as errors. I did not change these headers in that CL
because some build targets were not building properly. I've determined that the
problem was that these targets are missing a dependency in their BUILD.gn files.
This CL adds this dependency to those build files and finishes the codemod.

Note: I have not run git cl format to put the headers in the correct order
because I ran into problems with the original CL due to this breaking fragile
code.

BUG=724264

Review-Url: https://codereview.chromium.org/2892703002
Cr-Original-Commit-Position: refs/heads/master@{#473333}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 38a14eb81c041f7ae7bff36226318bca3d161e79