Remove references to inherited_data_ in ComputedStyle.
To allow changes to where fields are stored, ComputedStyle code should
not refer directly to a group, as that code will break when we change
groups. This patch removes references to inherited_data_ in ComputedStyle,
replacing with generated or handwritten getters. This patch does not
remove references within diffing functions as those will soon be
generated.

BUG=710938

Review-Url: https://codereview.chromium.org/2886293004
Cr-Original-Commit-Position: refs/heads/master@{#473433}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1b298605df3e00d8bb378fb2c17a8014d0d1c578