DevTools: fix popover's account for scrollbar size when measuring content
When calculating the size of a popover with SizeBehavior == MeasureContent,
positioning logic accounts for either a vertical scrollbar or a horizontal one,
but not both. This led to cases where GlassPane allocated enough width, but not
enough height for wide content (e.g. hovering over a long string when paused).

This CL adjusts the constrained popover width/height for scrollbar size in both
directions.

BUG=722899

Review-Url: https://codereview.chromium.org/2889693008
Cr-Original-Commit-Position: refs/heads/master@{#473499}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7856682b892241cb9263140c56554880633b8d84