Upstream service worker "script URL" test to WPT
This test is nearly identical to the version available in the Web
Platform Tests project. Update the upstream version with the only
meaningful difference between the tests: an additional assertion for the
interpretation of fragment identifiers. Re-factor test body to use
`promise_test` and format code accordingly. Add a "use strict"
directive. Remove the Chromium version of the test.

BUG=688116
R=falken@chromium.org

Review-Url: https://codereview.chromium.org/2894883003
Cr-Original-Commit-Position: refs/heads/master@{#473586}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a51bf4dc7726796505882c1b96eed0ec40c28d4f