Move CalculateScrollbarModes from FrameView to LayoutView.
This is a better place given that it is used by both FrameView and the root
PaintLayerScrollableArea.

SetCanHaveScrollbars remains on FrameView, but with a comment to clarify that it
is RLS-aware.

BUG=724309
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_layout_tests_slimming_paint_v2

Review-Url: https://codereview.chromium.org/2900513002
Cr-Original-Commit-Position: refs/heads/master@{#473695}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1c38965c165d21706028861200fab06b1eca0bea