net: Do not sniff XHTML content
This CL removes XHTML content sniff for text/xml and application/xml.

We altered the content-type to application/xhtml+xml if the root tag starts with
the following:

<html xmlns="http://www.w3.org/1999/xhtml"

We stop it.
This behavior has never been applied to

<html xmlns='http://www.w3.org/1999/xhtml'
or
<html  xmlns="http://www.w3.org/1999/xhtml"

because this behavior was introduced to pass some layout tests.

The behavior difference between application/xml and application/xhtml+xml is
document.createElement().  This CL updates some tests.
New behavior is interoperable with Edge, Firefox, and Safari.

* LayoutTests changes
** external/wpt/dom/nodes/Document-createElement-namespace.html
  This CL fixes failing sub-tests.
** fast/dom/Document/xml-document-focus.xml
  Page rendering is changed because of document.createElement() in js-test.js.
** fast/dom/ping-attribute-no-crash.xml
  This assumes document.createElement() has XHTML behavior. This CL renames it
  to html/text_level_semantics/a-ping-no-crash.xhtml.
** http/tests/misc/createElementNamespace1.xml
  This CL fixes a failing sub-test about document.createElement().

BUG=231927, 702084

Review-Url: https://codereview.chromium.org/2883833002
Cr-Original-Commit-Position: refs/heads/master@{#473805}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9cc9427a4b67b394747dde2f1b8279a0d3384892