Replace NinePieceImage.setMaskDefault with static method.
For NinePieceImages that are used as masks, we currently create a
default NinePieceImage and then call setMaskDefault on it. This is
unnecessarily complicated. Instead, we could just have a static method
on NinePieceImage that creates a mask NinePieceImage. This patch adds
a static method NinePieceImage::MaskDefaults() that returns a
NinePieceImage with its members initialised to mask defaults.

BUG=628043

Review-Url: https://codereview.chromium.org/2901433002
Cr-Original-Commit-Position: refs/heads/master@{#474177}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 43a17a2ec583e217c93c9895cfe90388b7c8ed2e