PlzNavigate: Fix missing console warning.
When PlzNavigate(--enable-browser-side-navigation) is enabled,
BaseFetchContext::CanRequestInternal([...]) is not used when the
resource to fetch is a frame. This implies that the code part that
check whether or not an url contains new lines in it was not used.
This CL duplicates this part into the FrameLoader.

BUG=723595
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_chromium_browser_side_navigation_rel

Review-Url: https://codereview.chromium.org/2898973002
Cr-Original-Commit-Position: refs/heads/master@{#474250}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9115376b363cf621a47d8c434c007b9c9e932b57