Ensure visual overflow from style recalc gets propagated
When recalculating overflow due to style change, the overflow we calculate may
be clipped. We still need to treat it as having changed so that we can update
the ContentsVisualOverflowRect() of our container. This will allow the paint layer
to clip the overflow when painting it.

BUG=724453

Review-Url: https://codereview.chromium.org/2896363003
Cr-Original-Commit-Position: refs/heads/master@{#474278}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eac1506ac1502a44f10c1af30d0c4de6d7490a6b