Generate enum/getters/setters/mappings for text-security.
Currently, the text-security property is a 'storage_only' field,
so it has no generated public getters/setters. This patch changes it to
a 'keyword' field so that its getters/setters can be generated as well.
We also generate the EWordBreak enum and use the generated
CSSValueID <-> EWordBreak mappings.

Diff of generated files:
https://gist.github.com/darrnshn/2811407420cd06be91eca37e52c974d6/revisions

BUG=628043

Review-Url: https://codereview.chromium.org/2911583002
Cr-Original-Commit-Position: refs/heads/master@{#474901}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dd404929f0f3d121a74361a394e1727fee2189a6