Fix bug in DocumentMarkerController::ShowMarkers()
I accidentally introduced a bug in DocumentMarkerController::ShowMarkers() in
https://codereview.chromium.org/2820633002 causing it to crash if we have a node
with a DocumentMarkerList for at least one MarkerType but not all MarkerTypes
(this is very common). I hadn't noticed it until just now because this is a
method only used to assist with debugging. This CL fixes the bug.

BUG=

Review-Url: https://codereview.chromium.org/2908593003
Cr-Original-Commit-Position: refs/heads/master@{#474910}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0cab0229765caa47ee02c482d768ea65f6e24b5d